maked refactoring
