+++ webapp/create_db.py
@@ -1,5 +1,4 @@
-import psycopg2
-from config import NAME, NAME_DB, PASSWORD
+from webapp.config import NAME, NAME_DB, PASSWORD
 from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import scoped_session, sessionmaker
@@ -10,4 +9,3 @@ db_session = scoped_session(sessionmaker(bind=engine))
 
 Base = declarative_base()
 Base.query = db_session.query_property()
-
