+++ webapp/templates/blank.html
@@ -11,25 +11,34 @@
   <link href="{{ url_for('static', filename='css/style.css') }}" type="text/css" rel="stylesheet" media="screen,projection"/>
 </head>
 <body style="background-color:LightGoldenRodYellow">
-  {% with messages = get_flashed_messages() %}
-            {% if messages %}
-          <div class="alert alert-info" role="alert">
+  {% with messages = get_flashed_messages() %} 
+      {% if messages %}
           {% for message in messages %}
-            {{ message }}<br>
+              <div class="card blue lighten-4"> 
+                  <div class="card-content">
+                      <p>{{ message }}</p> 
+                  </div>
+              </div>
           {% endfor %}
-          </div>
-          {% endif %}
-            {% endwith %}
+      {% endif %}
+  {% endwith %} 
   <nav class="green darken-1" role="navigation">
     <div class="nav-wrapper container"><a id="logo-container" href="/" class="brand-logo">LearnEnglish</a>
       <ul class="right hide-on-med-and-down">
-        <li><a href="/" style="font-size:125%">Главная</a></li>
-        <li><a href="/alphabet" style="font-size:125%">Алфавит</a></li>
-        <li><a href="/syllables" style="font-size:125%">Слоги</a></li>
-        <li><a href="/register" style="font-size:125%">Зарегистрироваться</a></li>
-        <li><a href="/login" style="font-size:125%">Войти</a></li>
-        <li><a href="/logout" style="font-size:125%">Выйти</a></li>
-        <li><a href="/admin" style="font-size:125%">Админка</a></li>
+<!-- Тут надо вставит проверку, если юзер не зареган, то показывать только кнопки "Зарег-ся" и "войти"  -->
+        <li><a href="/" style="font-size:100%">Главная</a></li>
+<!-- Тут надо вставит проверку, если юзер вошел, то показывать кнопку "Зарег-ся" не надо  -->
+        {% if current_user.is_authenticated %}
+        <li><a href="/content/alphabet" style="font-size:100%">Алфавит</a></li>
+        <li><a href="/content/syllables" style="font-size:100%">Слоги</a></li>
+        <li><a href="/users/logout" style="font-size:100%">Выйти</a></li>
+        {% else %}
+        <li><a href="/users/register" style="font-size:100%">Зарегистрироваться</a></li>
+        <li><a href="/users/login" style="font-size:100%">Войти</a></li>
+        {% endif %}
+        {% if current_user.is_admin %}
+        <li><a href="/admin" style="font-size:100%">Админка</a></li>
+        {% endif %}
       </ul>
       <ul id="nav-mobile" class="sidenav green darken-1">
         <br>
