+++ webapp/templates/login.html
@@ -2,70 +2,77 @@
 <!doctype html>
 <html lang="ru">
   <head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <!--Import Google Icon Font-->
+    <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+    <!--Import materialize.css-->
+    <link type="text/css" rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css"  media="screen,projection"/>
 
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.1.3/dist/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
-
-    <title>{{ page_title }}</title>
-  </head>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
   <body>
     <div class="container">
-    <h1>{{ page_title }}</h1>
+    <h3>{{ page_title }}</h3>
     <div class="row">
-        <div class="col-6">
-          {% with messages = get_flashed_messages() %}
-          {% if messages %}
-            <div class="card blue lighten-4"> 
-              <div class="card-content">
+      {% with messages = get_flashed_messages() %} 
+            {% if messages %}
                 {% for message in messages %}
-                  <p>{{ message }}</p>
+                    <div class="card blue lighten-4"> 
+                        <div class="card-content">
+                            <p>{{ message }}</p> 
+                        </div>
+                    </div>
                 {% endfor %}
+            {% endif %}
+        {% endwith %} 
+        {% if show_register %}
+          <div class="col-6">
+            <form action="" method="POST">
+                {{ register_form.hidden_tag() }}
+                <h5>Регистрация</h5>
+                <div class="input-field">
+                {{ register_form.user_name.label }}<br>{{ register_form.user_name() }}
+                </div>
+                <div class="input-field">
+                {{ register_form.nickname.label }}<br>{{ register_form.nickname() }}
+                </div>
+                <div class="input-field">
+                {{ register_form.email.label }}<br>{{ register_form.email() }}
+                </div>
+                <div class="input-field">
+                  {{ register_form.password.label }}<br>{{ register_form.password() }}
+                </div>
+                <div class="input-field">
+                  {{ register_form.reapet_password.label }}<br>{{ register_form.reapet_password() }}
+                </div>
+                {{ register_form.submit }}
+            </form>
+          </div>
+        {% else %}
+          <div class="col-6">
+            <form action="{{ url_for('users.process_login') }}" method="POST">
+              {{ login_form.hidden_tag() }}
+              <h5>Авторизация</h5>
+              <div class="input-field">
+              {{ login_form.nickname2.label }}<br>{{ login_form.nickname2() }}
               </div>
-            </div>
-          {% endif %}
-        {% endwith %}
-        
-        <form action="" method="POST">
-            {{ register_form.hidden_tag() }}
-            <h2>Регистрация</h2>
-            <div class="form-group">
-            {{ register_form.user_name.label }}<br>{{ register_form.user_name() }}
-            </div>
-            <div class="form-group">
-            {{ register_form.nickname.label }}<br>{{ register_form.nickname() }}
-            </div>
-            <div class="form-group">
-            {{ register_form.email.label }}<br>{{ register_form.email() }}
-            </div>
-            <div class="form-group">
-            {{ register_form.password.label }}<br>{{ register_form.password() }}
-            </div>
-            {{ register_form.submit }}
-        </form>
-        </div>
-        
-        <div class="col-6">
-          <form action="{{ url_for('process_login') }}" method="POST">
-            {{ login_form.hidden_tag() }}
-            <h2>Авторизация</h2>
-            <div class="form-group">
-            {{ login_form.nickname2.label }}<br>{{ login_form.nickname2() }}
-            </div>
-            <div class="form-group">
-            {{ login_form.password2.label }}<br>{{ login_form.password2() }}
-            </div>
+              <div class="input-field">
+              {{ login_form.password2.label }}<br>{{ login_form.password2() }}
+              </div>
+              <div>
+                <label>
+                  {{ login_form.remember_me(class_="filled-in", id="filled-in-box") }}
+                  <span class="black-text">{{ login_form.remember_me.label.text }}</span>
+              </label>
+              </div>
             {{ login_form.submit2 }}
-        </form>    
-        </div>
+            </form>    
+          </div>
+          {% endif %}
     </div>
     </div>
-    <!-- Optional JavaScript -->
-    <!-- jQuery first, then Popper.js, then Bootstrap JS -->
-    <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.14.3/dist/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.1.3/dist/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
+    <script type="text/javascript" src="https://code.jquery.com/jquery-2.1.1.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
   </body>
 </html>
+
+
+
