+++ webapp/user/models.py
@@ -24,9 +24,9 @@ class User(Base, UserMixin):
     def check_password(self, password):
         return check_password_hash(self.password, password)
 
-    # @property
-    # def is_admin(self):
-    #     return self.role == 'admin'
+    @property
+    def is_admin(self):
+        return self.role == 'admin'
 
     def __repr__(self):
         return f'<User {self.name} {self.nickname} {self.email} {self.role}>'
